SavedCase Int16 changed to Int64

+++ Covid19App/CoreData/Model/SavedCase+CoreDataProperties.swift
@@ -20,9 +20,9 @@ extension SavedCase {
     @NSManaged public var country: String?
     @NSManaged public var countryCode: String?
     @NSManaged public var date: String?
-    @NSManaged public var newConfirmed: Int16
-    @NSManaged public var newDeaths: Int16
-    @NSManaged public var newRecovered: Int16
+    @NSManaged public var newConfirmed: Int64
+    @NSManaged public var newDeaths: Int64
+    @NSManaged public var newRecovered: Int64
     @NSManaged public var slug: String?
     @NSManaged public var totalConfirmed: Int64
     @NSManaged public var totalDeaths: Int64
